Add Wrayth parser tests; split background tests into their own file
Move the <background> tag tests (and their backgroundEvent helper) out of
WraythParserTests into a dedicated WraythBackgroundTests, and expand
WraythParserTests with general protocol-parser coverage asserted against the
Content model: start/empty tags with attributes and text, valueless attributes
falling back to name-as-value, literal and entity-escaped angle brackets inside
attribute strings (verifying the lexer's STRING rule doesn't end the tag early),
embedded quotes in attribute values, entity/numeric character references in
text, and multiple sibling elements.
